Refactored imports. Embedded elasticsearch http is now disabled.
