+++ app/controllers/ListIndices.scala
@@ -1,15 +1,10 @@
 package controllers
 
 import _root_.helper.services.IndicesStatsService
-import _root_.helper.utils.{ElasticsearchClient, AuthenticatedAction}
+import _root_.helper.utils.AuthenticatedAction
 import play.api.mvc._
-import esclient.EsClient
-import esclient.queries.GetFillableIndicesQuery
-import java.net.ConnectException
-import models._
+import esclient.ElasticsearchClient
 import views._
-import play.api.i18n.Messages
-import org.elasticsearch.bootstrap.ElasticSearch
 
 object ListIndices extends Controller {
