+++ app/esclient/ElasticsearchClient.scala
@@ -1,4 +1,4 @@
-package helper.utils
+package esclient
 
 import org.elasticsearch.node.Node
 import org.elasticsearch.client.Client
@@ -8,7 +8,7 @@ import org.elasticsearch.common.settings.ImmutableSettings
 object ElasticsearchClient {
   val settings = ImmutableSettings.settingsBuilder()
     .put("path.data", "data")
-    .put("http.enabled", true)
+    .put("http.enabled", false)
   val node: Node = nodeBuilder().settings(settings).clusterName("fbl_cluster").data(true).node()
   val elasticClient: Client = node.client()
